refactor(worker-sessions): split lifecycle control surfaces
