+++ pkg/services/worker_sessions/internal/service/service_test.go
@@ -4,10 +4,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"strings"
 	"sync"
+	"sync/atomic"
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/platform/logging"
+	"github.com/portpowered/infinite-you/pkg/services/providers"
 	workersessions "github.com/portpowered/infinite-you/pkg/services/worker_sessions"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
 )
@@ -397,4 +400,526 @@ func equalStrings(a, b []string) bool {
 		}
 	}
 	return true
+}
+
+func TestInvokeSession_RetriesARetryableFailureUnderOneWorkerIdentity(t *testing.T) {
+	var attempts atomic.Int32
+	execution := &fakeExecution{
+		dispatch: func(_ context.Context, req workers.WorkstationDispatchRequest) (workers.WorkstationDispatchResult, error) {
+			if attempts.Add(1) == 1 {
+				return retryableFailureResult(req), nil
+			}
+			return acceptedResult(req), nil
+		},
+	}
+	registry := newRegistryWithExecution(execution)
+
+	req := validStartRequest("worker-1", "dispatch-1")
+	req.Retry = workersessions.RetryPolicy{MaxAttempts: 2}
+	result, err := registry.InvokeSession(context.Background(), req)
+	if err != nil {
+		t.Fatalf("InvokeSession: %v", err)
+	}
+	if result.Session.State != workersessions.StateCompleted {
+		t.Fatalf("session state = %q, want COMPLETED after the retry succeeded", result.Session.State)
+	}
+	if result.Attempts != 2 {
+		t.Fatalf("attempts = %d, want 2", result.Attempts)
+	}
+
+	requests := execution.requests()
+	if len(requests) != 2 {
+		t.Fatalf("Workers dispatches = %d, want 2", len(requests))
+	}
+	if got := requests[0].Execution.Dispatch.DispatchID; got != "dispatch-1" {
+		t.Fatalf("first attempt dispatch ID = %q, want the caller's own", got)
+	}
+	if got := requests[1].Execution.Dispatch.DispatchID; got != "dispatch-1/attempt/2" {
+		t.Fatalf("second attempt dispatch ID = %q, want dispatch-1/attempt/2", got)
+	}
+	if strings.TrimSpace(result.Session.ID) != "worker-1" {
+		t.Fatalf("session ID = %q, want the one Worker identity to survive the retry", result.Session.ID)
+	}
+}
+
+// TestInvokeSession_StopsAtTheAttemptBudget proves the budget is a ceiling on
+// attempts, not on retries after the first: a Worker allowed two attempts that
+// fails both is terminal, and Workers is not asked a third time.
+func TestInvokeSession_StopsAtTheAttemptBudget(t *testing.T) {
+	execution := &fakeExecution{
+		dispatch: func(_ context.Context, req workers.WorkstationDispatchRequest) (workers.WorkstationDispatchResult, error) {
+			return retryableFailureResult(req), nil
+		},
+	}
+	registry := newRegistryWithExecution(execution)
+
+	req := validStartRequest("worker-1", "dispatch-1")
+	req.Retry = workersessions.RetryPolicy{MaxAttempts: 2}
+	result, err := registry.InvokeSession(context.Background(), req)
+	if err != nil {
+		t.Fatalf("InvokeSession: %v", err)
+	}
+	if !result.Session.Terminal() {
+		t.Fatalf("session state = %q, want a terminal state once the budget is spent", result.Session.State)
+	}
+	if execution.callCount() != 2 {
+		t.Fatalf("Workers dispatches = %d, want exactly the 2-attempt budget", execution.callCount())
+	}
+	if result.Attempts != 2 {
+		t.Fatalf("attempts = %d, want 2", result.Attempts)
+	}
+}
+
+// TestInvokeSession_DefaultPolicyIsOneAttempt pins the property that lets one
+// operation serve both orchestrators. A Petri dispatch has always been one
+// attempt with retryability classified and handed outward for the graph to act
+// on; converging JavaScript children onto this call must not quietly give
+// every Petri Worker attempt-level retry it never had.
+func TestInvokeSession_DefaultPolicyIsOneAttempt(t *testing.T) {
+	execution := &fakeExecution{
+		dispatch: func(_ context.Context, req workers.WorkstationDispatchRequest) (workers.WorkstationDispatchResult, error) {
+			return retryableFailureResult(req), nil
+		},
+	}
+	registry := newRegistryWithExecution(execution)
+
+	result, err := registry.InvokeSession(context.Background(), validStartRequest("worker-1", "dispatch-1"))
+	if err != nil {
+		t.Fatalf("InvokeSession: %v", err)
+	}
+	if execution.callCount() != 1 {
+		t.Fatalf("Workers dispatches = %d, want 1 for the zero-value retry policy", execution.callCount())
+	}
+	if result.Attempts != 1 {
+		t.Fatalf("attempts = %d, want 1", result.Attempts)
+	}
+}
+
+// TestInvokeSession_DoesNotRetryATerminalFailure keeps the retry decision
+// Workers' own: a failure Workers classifies as terminal is terminal here too,
+// because a second opinion would let two orchestrators disagree about the
+// identical provider failure.
+func TestInvokeSession_DoesNotRetryATerminalFailure(t *testing.T) {
+	execution := &fakeExecution{
+		dispatch: func(_ context.Context, req workers.WorkstationDispatchRequest) (workers.WorkstationDispatchResult, error) {
+			result := retryableFailureResult(req)
+			result.Result.FailureMetadata = &workers.WorkFailureMetadata{
+				Family: workers.WorkFailureFamilyTerminal,
+				Type:   workers.WorkFailureTypePermanentBadRequest,
+			}
+			return result, nil
+		},
+	}
+	registry := newRegistryWithExecution(execution)
+
+	req := validStartRequest("worker-1", "dispatch-1")
+	req.Retry = workersessions.RetryPolicy{MaxAttempts: 5}
+	if _, err := registry.InvokeSession(context.Background(), req); err != nil {
+		t.Fatalf("InvokeSession: %v", err)
+	}
+	if execution.callCount() != 1 {
+		t.Fatalf("Workers dispatches = %d, want 1; a terminal classification is not retried", execution.callCount())
+	}
+}
+
+func retryableFailureResult(req workers.WorkstationDispatchRequest) workers.WorkstationDispatchResult {
+	dispatchID := req.Execution.Dispatch.DispatchID
+	return workers.WorkstationDispatchResult{
+		DispatchID:      dispatchID,
+		WorkstationName: req.WorkstationName,
+		TerminalOutcome: workers.WorkstationDispatchTerminalOutcomeFailed,
+		Result: workers.WorkResult{
+			DispatchID: dispatchID,
+			Outcome:    workers.OutcomeFailed,
+			FailureMetadata: &workers.WorkFailureMetadata{
+				Family: workers.WorkFailureFamilyRetryable,
+				Type:   workers.WorkFailureTypeInternalServerError,
+			},
+		},
+	}
+}
+
+func acceptedResult(req workers.WorkstationDispatchRequest) workers.WorkstationDispatchResult {
+	dispatchID := req.Execution.Dispatch.DispatchID
+	return workers.WorkstationDispatchResult{
+		DispatchID:      dispatchID,
+		WorkstationName: req.WorkstationName,
+		TerminalOutcome: workers.WorkstationDispatchTerminalOutcomeCompleted,
+		Result: workers.WorkResult{
+			DispatchID: dispatchID,
+			Outcome:    workers.OutcomeAccepted,
+		},
+	}
+}
+
+// TestInvokeSession_ControlDuringAnAttemptSpendsNoRetryBudget closes the race
+// between the attempt loop and a control.
+//
+// Every disqualifying condition is checked before the budget, so a Worker a
+// control already owns can never consume an attempt: publishing another would
+// run provider work for a session that has moved on.
+func TestInvokeSession_ControlDuringAnAttemptSpendsNoRetryBudget(t *testing.T) {
+	execution := newGatedRetryExecution()
+	registry := newRegistryWithExecution(execution)
+
+	req := validStartRequest("worker-1", "dispatch-1")
+	req.Retry = workersessions.RetryPolicy{MaxAttempts: 3}
+	done := make(chan workersessions.InvokeSessionResult, 1)
+	go func() {
+		result, err := registry.InvokeSession(context.Background(), req)
+		if err != nil {
+			t.Errorf("InvokeSession: %v", err)
+		}
+		done <- result
+	}()
+
+	<-execution.entered
+	if _, err := registry.Cancel(context.Background(), workersessions.ControlRequest{ID: "worker-1"}); err != nil {
+		t.Fatalf("Cancel: %v", err)
+	}
+
+	result := <-done
+	if !result.Session.Terminal() {
+		t.Fatalf("session state = %q, want terminal after the control won", result.Session.State)
+	}
+	if execution.callCount() != 1 {
+		t.Fatalf("Workers dispatches = %d, want 1; a canceled Worker must not consume its retry budget",
+			execution.callCount())
+	}
+}
+
+// gatedRetryExecution holds its first attempt open until a control cancels it,
+// so the retry decision is made for a Worker a control already owns.
+type gatedRetryExecution struct {
+	*fakeExecution
+
+	entered     chan struct{}
+	enteredOnce sync.Once
+	release     chan struct{}
+	releaseOnce sync.Once
+}
+
+func newGatedRetryExecution() *gatedRetryExecution {
+	gated := &gatedRetryExecution{
+		entered: make(chan struct{}),
+		release: make(chan struct{}),
+	}
+	gated.fakeExecution = &fakeExecution{
+		dispatch: func(ctx context.Context, req workers.WorkstationDispatchRequest) (workers.WorkstationDispatchResult, error) {
+			gated.enteredOnce.Do(func() { close(gated.entered) })
+			select {
+			case <-gated.release:
+			case <-ctx.Done():
+			}
+			return retryableFailureResult(req), nil
+		},
+		cancel: func(_ context.Context, req workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+			gated.releaseOnce.Do(func() { close(gated.release) })
+			return workers.WorkstationDispatchCancelResult{
+				DispatchID: req.DispatchID,
+				Outcome:    workers.WorkstationDispatchCancelOutcomeCanceled,
+			}, nil
+		},
+	}
+	return gated
+}
+
+func TestInterrupt_CancelsExactSourceBeforeAdmittingExactSessionSuccessor(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	sourceResult, reference := prepareInterruptSource(t, registry, boundary)
+
+	request := workersessions.InterruptRequest{
+		RequestID:                "interrupt-request-1",
+		SourceWorkerSessionID:    "source-session",
+		SuccessorWorkerSessionID: "successor-session",
+		ReplacementMessage:       " replacement message ",
+	}
+	result, err := registry.Interrupt(context.Background(), request)
+	if err != nil {
+		t.Fatalf("Interrupt() error = %v", err)
+	}
+	assertInterruptAcceptedResult(t, result, reference)
+	cancellations := boundary.cancellations()
+	if len(cancellations) != 1 || cancellations[0].DispatchID != "dispatch-source" {
+		t.Fatalf("cancellations = %#v, want one exact source dispatch", cancellations)
+	}
+
+	handoff := boundary.currentRequest()
+	assertInterruptHandoff(t, handoff, request, reference)
+	if sourceResult := <-sourceResult; sourceResult.Session.State != workersessions.StateCanceled {
+		t.Fatalf("source InvokeSession() = %#v, want CANCELED", sourceResult.Session)
+	}
+
+	boundary.complete(completedDispatchWithProviderSession(handoff.Execution.Dispatch.DispatchID, reference), nil)
+	finalSuccessor, err := registry.Get(context.Background(), workersessions.GetRequest{ID: "successor-session"})
+	if err != nil {
+		t.Fatalf("Get(successor) error = %v", err)
+	}
+	if finalSuccessor.State != workersessions.StateCompleted {
+		t.Fatalf("final successor = %#v, want COMPLETED", finalSuccessor)
+	}
+}
+
+func prepareInterruptSource(
+	t *testing.T,
+	registry workersessions.Service,
+	boundary *controlledBoundary,
+) (<-chan workersessions.InvokeSessionResult, providers.SessionRef) {
+	t.Helper()
+	sourceResult := startControlledSession(t, registry, boundary, "source-session", "dispatch-source")
+	reference := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-interrupt"}
+	if _, err := registry.AssociateProviderSession(context.Background(), workersessions.ProviderSessionAssociationRequest{
+		WorkerSessionID: "source-session", DispatchID: "dispatch-source", Reference: reference,
+	}); err != nil {
+		t.Fatalf("AssociateProviderSession() error = %v", err)
+	}
+	boundary.setCancel(func(_ context.Context, request workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+		if request.DispatchID != "dispatch-source" {
+			t.Errorf("cancel dispatch = %q, want dispatch-source", request.DispatchID)
+		}
+		if got := boundary.publishCount(); got != 1 {
+			t.Errorf("publish count during source cancellation = %d, want 1", got)
+		}
+		boundary.complete(canceledDispatchResult("dispatch-source"), workers.ErrWorkstationDispatchCanceled)
+		return workers.WorkstationDispatchCancelResult{DispatchID: "dispatch-source", Outcome: workers.WorkstationDispatchCancelOutcomeCanceled}, nil
+	})
+	return sourceResult, reference
+}
+
+func assertInterruptAcceptedResult(t *testing.T, result workersessions.InterruptResult, reference providers.SessionRef) {
+	t.Helper()
+	if !result.Accepted || result.Phase != workersessions.InterruptPhaseSuccessorAdmission {
+		t.Fatalf("Interrupt() result = %#v, want accepted successor admission", result)
+	}
+	if result.Source.State != workersessions.StateCanceled || result.Source.SuccessorWorkerSessionID != "successor-session" {
+		t.Fatalf("interrupt source = %#v, want CANCELED with successor lineage", result.Source)
+	}
+	if result.Successor.State != workersessions.StateRunning || result.Successor.PredecessorWorkerSessionID != "source-session" {
+		t.Fatalf("interrupt successor = %#v, want RUNNING with predecessor lineage", result.Successor)
+	}
+	if result.Source.ProviderSessionAssociation == nil || result.Source.ProviderSessionAssociation.Reference != reference {
+		t.Fatalf("interrupt source reference = %#v, want exact %v", result.Source.ProviderSessionAssociation, reference)
+	}
+	if result.Successor.ProviderSessionAssociation == nil || result.Successor.ProviderSessionAssociation.Reference != reference {
+		t.Fatalf("interrupt successor reference = %#v, want exact %v", result.Successor.ProviderSessionAssociation, reference)
+	}
+}
+
+func assertInterruptHandoff(
+	t *testing.T,
+	handoff workers.WorkstationDispatchRequest,
+	request workersessions.InterruptRequest,
+	reference providers.SessionRef,
+) {
+	t.Helper()
+	if handoff.Execution.UserMessage != request.ReplacementMessage {
+		t.Fatalf("replacement message = %q, want byte-equivalent %q", handoff.Execution.UserMessage, request.ReplacementMessage)
+	}
+	if handoff.Execution.ResumeSession == nil || *handoff.Execution.ResumeSession != reference {
+		t.Fatalf("successor ResumeSession = %#v, want exact %v", handoff.Execution.ResumeSession, reference)
+	}
+	if handoff.Execution.Dispatch.DispatchID == "dispatch-source" || handoff.Execution.Dispatch.DispatchID == "" {
+		t.Fatalf("successor dispatch ID = %q, want distinct non-empty ID", handoff.Execution.Dispatch.DispatchID)
+	}
+}
+
+func TestInterrupt_ValidationRejectsBeforeDispatchEffects(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	for name, request := range map[string]workersessions.InterruptRequest{
+		"missing request ID": {
+			SourceWorkerSessionID: "source-session", SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "replacement",
+		},
+		"same source and successor": {
+			RequestID: "interrupt-invalid", SourceWorkerSessionID: "source-session", SuccessorWorkerSessionID: "source-session", ReplacementMessage: "replacement",
+		},
+		"blank replacement": {
+			RequestID: "interrupt-invalid", SourceWorkerSessionID: "source-session", SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "  \t",
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			result, err := registry.Interrupt(context.Background(), request)
+			var interruptErr *workersessions.InterruptError
+			if !errors.As(err, &interruptErr) || interruptErr.Phase != workersessions.InterruptPhaseValidation ||
+				!errors.Is(err, workersessions.ErrInterruptValidation) || result.Phase != workersessions.InterruptPhaseValidation {
+				t.Fatalf("validation Interrupt() = %#v, %v, want VALIDATION", result, err)
+			}
+		})
+	}
+	if boundary.publishCount() != 0 || len(boundary.cancellations()) != 0 {
+		t.Fatalf("validation effects = publishes %d, cancels %d, want 0/0", boundary.publishCount(), len(boundary.cancellations()))
+	}
+}
+
+func TestInterrupt_IdempotencyAndRequestConflictAvoidDuplicateEffects(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	sourceResult := startControlledSession(t, registry, boundary, "source-session", "dispatch-source")
+	reference := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-idempotent"}
+	if _, err := registry.AssociateProviderSession(context.Background(), workersessions.ProviderSessionAssociationRequest{
+		WorkerSessionID: "source-session", DispatchID: "dispatch-source", Reference: reference,
+	}); err != nil {
+		t.Fatalf("AssociateProviderSession() error = %v", err)
+	}
+	boundary.setCancel(func(_ context.Context, request workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+		boundary.complete(canceledDispatchResult(request.DispatchID), workers.ErrWorkstationDispatchCanceled)
+		return workers.WorkstationDispatchCancelResult{DispatchID: request.DispatchID, Outcome: workers.WorkstationDispatchCancelOutcomeCanceled}, nil
+	})
+	request := workersessions.InterruptRequest{
+		RequestID: "interrupt-request-idempotent", SourceWorkerSessionID: "source-session",
+		SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "first replacement",
+	}
+	first, err := registry.Interrupt(context.Background(), request)
+	if err != nil || !first.Accepted {
+		t.Fatalf("first Interrupt() = %#v, %v, want accepted", first, err)
+	}
+	firstPublishCount := boundary.publishCount()
+	firstCancellationCount := len(boundary.cancellations())
+	second, err := registry.Interrupt(context.Background(), request)
+	if err != nil || !second.Accepted {
+		t.Fatalf("replayed Interrupt() = %#v, %v, want accepted replay", second, err)
+	}
+	if boundary.publishCount() != firstPublishCount || len(boundary.cancellations()) != firstCancellationCount {
+		t.Fatalf("replay effects changed: publishes=%d cancels=%d, want %d/%d", boundary.publishCount(), len(boundary.cancellations()), firstPublishCount, firstCancellationCount)
+	}
+	conflict := request
+	conflict.ReplacementMessage = "different replacement"
+	conflictResult, err := registry.Interrupt(context.Background(), conflict)
+	var interruptErr *workersessions.InterruptError
+	if !errors.As(err, &interruptErr) || !errors.Is(err, workersessions.ErrInterruptValidation) ||
+		!errors.Is(err, workersessions.ErrInterruptRequestIDConflict) || conflictResult.Phase != workersessions.InterruptPhaseValidation {
+		t.Fatalf("conflicting Interrupt() = %#v, %v, want validation/idempotency conflict", conflictResult, err)
+	}
+	if boundary.publishCount() != firstPublishCount || len(boundary.cancellations()) != firstCancellationCount {
+		t.Fatalf("conflict effects changed: publishes=%d cancels=%d", boundary.publishCount(), len(boundary.cancellations()))
+	}
+	handoff := boundary.currentRequest()
+	boundary.complete(completedDispatchWithProviderSession(handoff.Execution.Dispatch.DispatchID, reference), nil)
+	if got := <-sourceResult; got.Session.State != workersessions.StateCanceled {
+		t.Fatalf("source InvokeSession() = %#v, want CANCELED", got.Session)
+	}
+}
+
+func TestInterrupt_ReportsSourceCancellationPhaseWithoutSuccessor(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	sourceResult := startControlledSession(t, registry, boundary, "source-session", "dispatch-source")
+	reference := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-cancel-failure"}
+	if _, err := registry.AssociateProviderSession(context.Background(), workersessions.ProviderSessionAssociationRequest{
+		WorkerSessionID: "source-session", DispatchID: "dispatch-source", Reference: reference,
+	}); err != nil {
+		t.Fatalf("AssociateProviderSession() error = %v", err)
+	}
+	boundary.setCancel(func(context.Context, workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+		return workers.WorkstationDispatchCancelResult{}, errors.New("cancel boundary unavailable")
+	})
+	result, err := registry.Interrupt(context.Background(), workersessions.InterruptRequest{
+		RequestID: "interrupt-cancel-failure", SourceWorkerSessionID: "source-session",
+		SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "replacement",
+	})
+	var interruptErr *workersessions.InterruptError
+	if !errors.As(err, &interruptErr) || interruptErr.Phase != workersessions.InterruptPhaseSourceCancellation ||
+		!errors.Is(err, workersessions.ErrInterruptSourceCancellation) || result.Phase != workersessions.InterruptPhaseSourceCancellation ||
+		result.Source.State != workersessions.StateRunning || result.Successor.ID != "" {
+		t.Fatalf("source cancellation failure = %#v, %v, want SOURCE_CANCELLATION with running source", result, err)
+	}
+	if boundary.publishCount() != 1 || len(boundary.cancellations()) != 1 {
+		t.Fatalf("source cancellation failure effects = publishes %d, cancels %d, want 1/1", boundary.publishCount(), len(boundary.cancellations()))
+	}
+	boundary.setCancel(nil)
+	boundary.complete(completedDispatchResult("dispatch-source"), nil)
+	if got := <-sourceResult; got.Session.State != workersessions.StateCompleted {
+		t.Fatalf("source cleanup InvokeSession() = %#v, want COMPLETED", got.Session)
+	}
+}
+
+func TestInterrupt_ReportsSuccessorAdmissionPhaseAfterSourceCancellation(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	sourceResult := startControlledSession(t, registry, boundary, "source-session", "dispatch-source")
+	reference := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-admission-failure"}
+	if _, err := registry.AssociateProviderSession(context.Background(), workersessions.ProviderSessionAssociationRequest{
+		WorkerSessionID: "source-session", DispatchID: "dispatch-source", Reference: reference,
+	}); err != nil {
+		t.Fatalf("AssociateProviderSession() error = %v", err)
+	}
+	boundary.setCancel(func(_ context.Context, request workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+		boundary.complete(canceledDispatchResult(request.DispatchID), workers.ErrWorkstationDispatchCanceled)
+		return workers.WorkstationDispatchCancelResult{DispatchID: request.DispatchID, Outcome: workers.WorkstationDispatchCancelOutcomeCanceled}, nil
+	})
+	boundary.setPublishError(func(call int, _ workers.WorkstationDispatchRequest) error {
+		if call == 2 {
+			return errors.New("successor admission unavailable")
+		}
+		return nil
+	})
+	result, err := registry.Interrupt(context.Background(), workersessions.InterruptRequest{
+		RequestID: "interrupt-admission-failure", SourceWorkerSessionID: "source-session",
+		SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "replacement",
+	})
+	var interruptErr *workersessions.InterruptError
+	if !errors.As(err, &interruptErr) || interruptErr.Phase != workersessions.InterruptPhaseSuccessorAdmission ||
+		!errors.Is(err, workersessions.ErrInterruptSuccessorAdmission) || result.Phase != workersessions.InterruptPhaseSuccessorAdmission ||
+		result.Source.State != workersessions.StateCanceled || result.Successor.State != workersessions.StateFailed {
+		t.Fatalf("successor admission failure = %#v, %v, want SUCCESSOR_ADMISSION with canceled source and failed successor", result, err)
+	}
+	if boundary.publishCount() != 2 || len(boundary.cancellations()) != 1 {
+		t.Fatalf("successor admission failure effects = publishes %d, cancels %d, want 2/1", boundary.publishCount(), len(boundary.cancellations()))
+	}
+	if got := <-sourceResult; got.Session.State != workersessions.StateCanceled {
+		t.Fatalf("source cleanup InvokeSession() = %#v, want CANCELED", got.Session)
+	}
+}
+
+func TestInterrupt_ConcurrentIdenticalRequestsReplayOneOperation(t *testing.T) {
+	boundary := newControlledBoundary()
+	registry := newControlledRegistry(t, boundary)
+	sourceResult := startControlledSession(t, registry, boundary, "source-session", "dispatch-source")
+	reference := providers.SessionRef{Provider: providers.IDCodex, Kind: providers.SessionIDKind, ID: "provider-session-concurrent"}
+	if _, err := registry.AssociateProviderSession(context.Background(), workersessions.ProviderSessionAssociationRequest{
+		WorkerSessionID: "source-session", DispatchID: "dispatch-source", Reference: reference,
+	}); err != nil {
+		t.Fatalf("AssociateProviderSession() error = %v", err)
+	}
+	boundary.setCancel(func(_ context.Context, request workers.WorkstationDispatchCancelRequest) (workers.WorkstationDispatchCancelResult, error) {
+		boundary.complete(canceledDispatchResult(request.DispatchID), workers.ErrWorkstationDispatchCanceled)
+		return workers.WorkstationDispatchCancelResult{DispatchID: request.DispatchID, Outcome: workers.WorkstationDispatchCancelOutcomeCanceled}, nil
+	})
+	request := workersessions.InterruptRequest{
+		RequestID: "interrupt-concurrent", SourceWorkerSessionID: "source-session",
+		SuccessorWorkerSessionID: "successor-session", ReplacementMessage: "replacement",
+	}
+	type outcome struct {
+		result workersessions.InterruptResult
+		err    error
+	}
+	outcomes := make(chan outcome, 2)
+	var group sync.WaitGroup
+	group.Add(2)
+	for range 2 {
+		go func() {
+			defer group.Done()
+			result, err := registry.Interrupt(context.Background(), request)
+			outcomes <- outcome{result: result, err: err}
+		}()
+	}
+	group.Wait()
+	close(outcomes)
+	for result := range outcomes {
+		if result.err != nil || !result.result.Accepted {
+			t.Fatalf("concurrent Interrupt() = %#v, %v, want accepted replay", result.result, result.err)
+		}
+	}
+	if got := len(boundary.cancellations()); got != 1 {
+		t.Fatalf("concurrent interrupt cancellations = %d, want 1", got)
+	}
+	if got := boundary.publishCount(); got != 2 {
+		t.Fatalf("concurrent interrupt publishes = %d, want source plus one successor", got)
+	}
+	handoff := boundary.currentRequest()
+	boundary.complete(completedDispatchWithProviderSession(handoff.Execution.Dispatch.DispatchID, reference), nil)
+	if got := <-sourceResult; got.Session.State != workersessions.StateCanceled {
+		t.Fatalf("source cleanup InvokeSession() = %#v, want CANCELED", got.Session)
+	}
 }
